Added new centralized routing method

+++ internal/measure/pathping/client.go
@@ -21,10 +21,10 @@ func (c *Client) Measure(path []int) (Result, error) {
 	if pathPingServer == nil {
 		return Result{}, fmt.Errorf("pathping server not started")
 	}
-	return pathPingServer.Measure(func(path []int) []uint32 {
-		p := make([]uint32, len(path))
+	return pathPingServer.Measure(func(path []int) []uint16 {
+		p := make([]uint16, len(path))
 		for i, v := range path {
-			p[i] = uint32(v)
+			p[i] = uint16(v)
 		}
 		return p
 	}(path), MeasurementOptions{
